Added fpga2sdram interface. ** READ CAREFULLY
To get this to work you need to clock the fpga-to-hps SDRAM PLL clock on the GHRD project, compile and upload the new preloader to the microSD.
It's is recommended too to load the microSD with the GHRD project rbf file, because the interface get activated at start. Then you can load any sof file that use the same fpga2sdram interface (AVALON-MM 256bits).